feat: create root directory on first mount
* refactor deserialize_from

* create root dir inode

* encapsulate serialization in objects

* move constants to own file

* fix import errors

* refactor File to Disk

* fix bug when reading bitmap

* allocate data blocks for root directory

* refactor saving data to blocks to separate method

* clean up the init method

* remove unnecessary loop

* load disk on program startup

* update design doc

+++ src/main.rs
@@ -1,12 +1,32 @@
 mod mkfs;
 mod tiny;
 
-use tiny::TinyFS;
+use std::{fs::OpenOptions, path::Path};
+
+use tiny::{constants::Disk, TinyFS};
 
 fn main() {
-    mkfs::make("./tiny.img");
+    let path = Path::new("./tiny.img");
+    if !path.exists() {
+        mkfs::make("./tiny.img");
+    }
 
     let mount_path = "/tmp/tiny";
 
-    fuse::mount(TinyFS, &mount_path, &[]).expect("expected filesytem to mount");
+    fuse::mount(
+        TinyFS {
+            disk: load_disk(path),
+        },
+        &mount_path,
+        &[],
+    )
+    .expect("expected filesytem to mount");
+}
+
+fn load_disk(path: &Path) -> Disk {
+    OpenOptions::new()
+        .write(true)
+        .read(true)
+        .open(path)
+        .expect("file to have been opened")
 }
